+++ test/test_structured_2d.jl
@@ -182,8 +182,9 @@ isdir(outdir) && rm(outdir, recursive=true)
 
   @trixi_testset "elixir_euler_rayleigh_taylor_instability.jl" begin
     @test_trixi_include(joinpath(EXAMPLES_DIR, "elixir_euler_rayleigh_taylor_instability.jl"),
-      l2   = [0.07623680588263236, 0.007462660867185416, 0.0035846371501533103, 0.01074438328884887],
-      linf = [1.1657734093890613, 0.043256773855258314, 0.049225387018938464, 0.03815681528560022],
+      l2   = [0.06365630381017849, 0.007166887387738937, 0.002878708825497772, 0.010247678114070121],
+      linf = [0.4799214336153155, 0.024595483032220266, 0.02059808120543466, 0.03190756362943725],
+      cells_per_dimension = (8,8),
       tspan = (0.0, 0.3))
   end